fix(reference_generator): Infinite while loop
Do not try to lower the velocity in order to reach the next subgoal,
this resulted in an infinite while loop. Just try to reacht the next
subgoal, if this cannot be done, notify the caller. I will test whether
this performs well in practice, if not, I will try a finite optimization
with a max iterations.
